update person service

+++ src/main/java/com/example/springbootdemo/api/entity/Person.java
@@ -8,12 +8,11 @@ import java.io.Serializable;
 
 @Data
 @Entity
-//@Table(name = "person", schema = "public")
 public class Person implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
-    private Long id;
+    private Integer id;
     private String name;
     private String surname;
   //  private String email;
